# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateDeviceRequest:
    properties:
      device_uid:
        description: remoteId/MAC stable identifier
        type: string
      fw_rev:
        type: string
      hw_rev:
        type: string
      model:
        type: string
      serial_number:
        type: string
      vendor:
        type: string
    type: object
...
  handlers.DailyActivityRequest:
    properties:
      calories:
        type: integer
      device_date:
        type: string
      running_steps:
        type: integer
      sleep_duration_s:
        type: integer
      source:
        type: string
      sport_duration_s:
        type: integer
      total_steps:
        type: integer
      tz_offset_min:
        type: integer
      walk_distance_m:
        type: integer
    type: object
  handlers.DeviceEventPoint:
    properties:
      event_type:
        type: string
      payload:
        additionalProperties: {}
        type: object
      source:
        type: string
      ts_utc:
        type: string
    type: object
...
  handlers.IngestDeviceEventsRequest:
    properties:
      events:
        items:
          $ref: '#/definitions/handlers.DeviceEventPoint'
        type: array
    type: object
...
  handlers.StatusResponse:
    properties:
      status:
        example: deleted
        type: string
    type: object
...
  handlers.UpdateDeviceRequest:
    properties:
      fw_rev:
        type: string
      hw_rev:
        type: string
      last_seen_at:
        description: RFC3339
        type: string
      model:
        type: string
      serial_number:
        type: string
      vendor:
        type: string
    type: object
...
  handlers.UpsertResponse:
    properties:
      upserted:
        example: 1
        type: integer
    type: object
...
  httpx.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  store.DailyActivityResponse:
    properties:
      calories:
        type: integer
      device_date:
        type: string
      device_id:
        type: string
      running_steps:
        type: integer
      sleep_duration_s:
        type: integer
      source:
        type: string
      sport_duration_s:
        type: integer
      synced_at:
        type: string
      total_steps:
        type: integer
      tz_offset_min:
        type: integer
      walk_distance_m:
        type: integer
    type: object
  store.DeviceEventResponse:
    properties:
      device_id:
        type: string
      event_type:
        type: string
      payload:
        additionalProperties: {}
        type: object
      source:
        type: string
      synced_at:
        type: string
      ts_utc:
        type: string
    type: object
  store.DeviceRow:
    properties:
      created_at:
        type: string
      device_uid:
        type: string
      first_seen_at:
        type: string
      fw_rev:
        type: string
      hw_rev:
        type: string
      id:
        type: string
      last_seen_at:
        type: string
      model:
        type: string
      serial_number:
        type: string
      updated_at:
        type: string
      vendor:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for AGEM wearable users, devices, qring sensor data, sleep,
    and workouts.
  title: AGEM Backend API
  version: "1.0"
paths:
...
  /v1/devices/:
    post:
      consumes:
      - application/json
      description: Creates a wearable device registry record.
      parameters:
      - description: Device payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.DeviceRow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Create device
      tags:
      - Devices
  /v1/devices/{device_id}/activity/daily:
    get:
      description: Lists device-local daily activity summaries in an inclusive date
        range.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Inclusive device-local start date
        format: date
        in: query
        name: from
        required: true
        type: string
      - description: Inclusive device-local end date
        format: date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.DailyActivityResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List daily activity
      tags:
      - Daily Activity
    post:
      consumes:
      - application/json
      description: Creates or updates one device-local daily activity summary for
        a device.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Daily activity payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DailyActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.DailyActivityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Upsert daily activity
      tags:
      - Daily Activity
  /v1/devices/{device_id}/events:
    get:
      description: Lists device events in an RFC3339 time range, optionally filtered
        by event type.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Event type filter
        in: query
        name: event_type
        type: string
      - description: Inclusive RFC3339 start timestamp
        format: date-time
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive RFC3339 end timestamp
        format: date-time
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.DeviceEventResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List device events
      tags:
      - Device Events
    post:
      consumes:
      - application/json
      description: Idempotently ingests device events such as sync markers, battery
        state, binding events, or SDK notifications.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Device event payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IngestDeviceEventsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpsertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Ingest device events
      tags:
      - Device Events
...
  /v1/devices/{id}:
    delete:
      description: Deletes a device by API ID.
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Delete device
      tags:
      - Devices
    get:
      description: Returns one device by API ID.
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.DeviceRow'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get device
      tags:
      - Devices
    patch:
      consumes:
      - application/json
      description: Updates device metadata and last seen timestamp.
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Device update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.DeviceRow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Update device
      tags:
      - Devices
  /v1/devices/by_uid/{device_uid}:
    get:
      description: Returns one device by stable wearable UID or MAC address.
      parameters:
      - description: Stable wearable UID
        in: path
        name: device_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.DeviceRow'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get device by UID
      tags:
      - Devices
...
swagger: "2.0"